refactor(day-diet): move use cases and infrastructure files
