+++ src/sections/day-diet/components/CopyLastDayButton.tsx
@@ -4,7 +4,7 @@ import {
   fetchPreviousDayDiets,
   insertDayDiet,
   updateDayDiet,
-} from '~/modules/diet/day-diet/application/dayDiet'
+} from '~/modules/diet/day-diet/application/dayDietUseCases'
 import { currentDayDiet } from '~/modules/diet/day-diet/application/dayDietStore'
 import {
   createNewDayDiet,
